requirements.txt
beautifulsoup4==4.14.3
Flask==3.1.3
fpdf2==2.8.7
groq==1.1.1
gunicorn==25.1.0
lxml==6.0.2
pandas==3.0.1
pillow==12.1.1
python-docx==1.2.0
requests==2.32.5
beautifulsoup4
Flask
fpdf2
groq
gunicorn
lxml
pandas
pillow
python-docx
requests
beautifulsoup4
Flask
fpdf2
groq
gunicorn
lxml
pandas
pillow
python-docx
requests
